Split spotify functions
