+++ core/internals.py
@@ -117,15 +117,6 @@ def sanitize_title(title):
     return title
 
 
-def generate_token():
-    """ Generate the token. Please respect these credentials :) """
-    credentials = oauth2.SpotifyClientCredentials(
-        client_id='4fe3fecfe5334023a1472516cc99d805',
-        client_secret='0f02b7c483c04257984695007a4a8d5c')
-    token = credentials.get_access_token()
-    return token
-
-
 def filter_path(path):
     if not os.path.exists(path):
         os.makedirs(path)
